requirements.txt
sentence-transformers==2.2.2
Pillow==10.0.0
scikit-learn
numpy
streamlit
torch
fastapi 
uvicorn 
python-multipart 
faiss-cpu 
requests
pinecone
google-search-results
cmake
dlib
ultralytics
face_recognition